# docs/openapi.yaml
openapi: 3.0.3
info:
  title: apple-bridge
  description: macOS EventKit HTTP API — access Apple Reminders and Calendars
  version: 1.0.0
servers:
  - url: http://127.0.0.1:23487
    description: Local server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key created via `apple-bridge keys create`
  schemas:
    AlarmInput:
      type: object
      properties:
        relativeOffset:
          type: number
          description: Seconds relative to event/due date (negative = before)
          example: -3600
        absoluteDate:
          type: string
          format: date-time
          example: "2026-04-15T09:00:00Z"
    AlarmResponse:
      type: object
      properties:
        relativeOffset:
          type: number
          nullable: true
        absoluteDate:
          type: string
          format: date-time
          nullable: true
    RecurrenceRuleInput:
      type: object
      required: [frequency, interval]
      properties:
        frequency:
          type: string
          enum: [daily, weekly, monthly, yearly]
        interval:
          type: integer
          example: 1
        daysOfTheWeek:
          type: array
          items:
            type: string
            enum: [MO, TU, WE, TH, FR, SA, SU]
        daysOfTheMonth:
          type: array
          items:
            type: integer
        monthsOfTheYear:
          type: array
          items:
            type: integer
        end:
          type: object
          properties:
            date:
              type: string
              format: date-time
            count:
              type: integer
    RecurrenceRuleResponse:
      type: object
      properties:
        frequency:
          type: string
        interval:
          type: integer
        daysOfTheWeek:
          type: array
          items:
            type: string
          nullable: true
        daysOfTheMonth:
          type: array
          items:
            type: integer
          nullable: true
        monthsOfTheYear:
          type: array
          items:
            type: integer
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
        endCount:
          type: integer
          nullable: true
...
    CreateReminderListRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          example: Shopping
        color:
          type: string
          example: "#FF6B6B"
    UpdateReminderListRequest:
      type: object
      properties:
        title:
          type: string
        color:
          type: string
    ReminderListResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        color:
          type: string
          nullable: true
        isDefault:
          type: boolean
        sourceTitle:
          type: string
    CreateCalendarEventRequest:
      type: object
      required: [title, startDate, endDate]
      properties:
        title:
          type: string
          example: Team Standup
        notes:
          type: string
        calendarId:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        isAllDay:
          type: boolean
        location:
          type: string
        url:
          type: string
        availability:
          type: string
          enum: [busy, free, tentative, unavailable]
        alarms:
          type: array
          items:
            $ref: '#/components/schemas/AlarmInput'
        recurrenceRules:
          type: array
          items:
            $ref: '#/components/schemas/RecurrenceRuleInput'
    UpdateCalendarEventRequest:
      type: object
      properties:
        title:
          type: string
        notes:
          type: string
        calendarId:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        isAllDay:
          type: boolean
        location:
          type: string
        url:
          type: string
        availability:
          type: string
          enum: [busy, free, tentative, unavailable]
        alarms:
          type: array
          items:
            $ref: '#/components/schemas/AlarmInput'
        recurrenceRules:
          type: array
          items:
            $ref: '#/components/schemas/RecurrenceRuleInput'
    CalendarEventResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        notes:
          type: string
          nullable: true
        calendarId:
          type: string
        calendarTitle:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        isAllDay:
          type: boolean
        location:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
        availability:
          type: string
        status:
          type: string
        alarms:
          type: array
          items:
            $ref: '#/components/schemas/AlarmResponse'
        recurrenceRules:
          type: array
          items:
            $ref: '#/components/schemas/RecurrenceRuleResponse'
        attendees:
          type: array
          items:
            $ref: '#/components/schemas/AttendeeResponse'
        hasRecurrenceRules:
          type: boolean
        organizer:
          type: string
          nullable: true
        creationDate:
          type: string
          format: date-time
          nullable: true
        lastModifiedDate:
          type: string
          format: date-time
          nullable: true
    AttendeeResponse:
      type: object
      properties:
        name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        status:
          type: string
    CreateCalendarRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          example: Work
        color:
          type: string
          example: "#4A90D9"
    UpdateCalendarRequest:
      type: object
      properties:
        title:
          type: string
        color:
          type: string
    CalendarResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        color:
          type: string
          nullable: true
        type:
          type: string
        isDefault:
          type: boolean
        allowsContentModifications:
          type: boolean
        sourceTitle:
          type: string
...
paths:
...
  /reminder-lists:
    get:
      summary: List all reminder lists
      tags: [Reminder Lists]
      responses:
        '200':
          description: List of reminder lists
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReminderListResponse'
                  count:
                    type: integer
    post:
      summary: Create a reminder list
      tags: [Reminder Lists]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReminderListRequest'
      responses:
        '201':
          description: Reminder list created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderListResponse'
  /reminder-lists/{id}:
    get:
      summary: Get a reminder list
      tags: [Reminder Lists]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reminder list found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderListResponse'
        '404':
          description: Not found
    put:
      summary: Update a reminder list
      tags: [Reminder Lists]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReminderListRequest'
      responses:
        '200':
          description: Reminder list updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderListResponse'
        '404':
          description: Not found
    delete:
      summary: Delete a reminder list
      tags: [Reminder Lists]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
  /events:
    get:
      summary: List calendar events
      tags: [Calendar Events]
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: calendarId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CalendarEventResponse'
                  count:
                    type: integer
        '400':
          description: Missing required date parameters
    post:
      summary: Create a calendar event
      tags: [Calendar Events]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCalendarEventRequest'
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarEventResponse'
  /events/{id}:
    get:
      summary: Get a calendar event
      tags: [Calendar Events]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarEventResponse'
        '404':
          description: Not found
    put:
      summary: Update a calendar event (full)
      tags: [Calendar Events]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCalendarEventRequest'
      responses:
        '200':
          description: Event updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarEventResponse'
        '404':
          description: Not found
    patch:
      summary: Update a calendar event (partial)
      tags: [Calendar Events]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCalendarEventRequest'
      responses:
        '200':
          description: Event updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarEventResponse'
        '404':
          description: Not found
    delete:
      summary: Delete a calendar event
      tags: [Calendar Events]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: span
          in: query
          description: "\"this\" (default) or \"future\" for recurring events"
          schema:
            type: string
            enum: [this, future]
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
  /calendars:
    get:
      summary: List all event calendars
      tags: [Calendars]
      responses:
        '200':
          description: List of calendars
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CalendarResponse'
                  count:
                    type: integer
    post:
      summary: Create a calendar
      tags: [Calendars]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCalendarRequest'
      responses:
        '201':
          description: Calendar created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarResponse'
  /calendars/{id}:
    get:
      summary: Get a calendar
      tags: [Calendars]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Calendar found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarResponse'
        '404':
          description: Not found
    put:
      summary: Update a calendar
      tags: [Calendars]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCalendarRequest'
      responses:
        '200':
          description: Calendar updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarResponse'
        '404':
          description: Not found
    delete:
      summary: Delete a calendar
      tags: [Calendars]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found